# client/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: fbstachura@gmail.com
  description: StampWallet API Server REST Specification
  title: StampWallet API Server
  version: 0.1.0
servers:
- url: /
paths:
...
  /business/account:
    post:
      description: "This endpoint is used to attach a new business account to an existing,\
        \ logged in user account. Busies details are provided in the request. Responds\
        \ with business id and ids of banner and icon image slots."
      operationId: createBusinessAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostBusinessAccountRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostBusinessAccountResponse'
          description: Successful operation
        "400":
          content:
            application/json:
              example:
                status: INVALID_REQUEST
              schema:
                $ref: '#/components/schemas/DefaultResponse'
          description: Invalid request format
        "409":
          content:
            application/json:
              example:
                status: ALREADY_EXISTS
              schema:
                $ref: '#/components/schemas/DefaultResponse'
          description: Business already exists
        "401":
          content:
            application/json:
              example:
                status: UNAUTHORIZED
              schema:
                $ref: '#/components/schemas/DefaultResponse'
          description: Invalid credentials
      security:
      - sessionToken: []
      summary: Create a business account
      tags:
      - business
  /business/info:
    get:
      description: Responds with information about business owned by the logged in
        user.
      operationId: getBusinessAccountInfo
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetBusinessAccountResponse'
          description: Successful operation
        "401":
          content:
            application/json:
              example:
                status: UNAUTHORIZED
              schema:
                $ref: '#/components/schemas/DefaultResponse'
          description: Invalid credentials
        "404":
          content:
            application/json:
              example:
                status: NOT_FOUND
              schema:
                $ref: '#/components/schemas/DefaultResponse'
          description: Does not exist or forbidden
      security:
      - sessionToken: []
      summary: Get business info
      tags:
      - business
    patch:
      description: This endpoint is used to update business account data
      operationId: updateBusinessAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchBusinessAccountRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              example:
                status: OK
              schema:
                $ref: '#/components/schemas/DefaultResponse'
          description: Successful operation
        "400":
          content:
            application/json:
              example:
                status: INVALID_REQUEST
              schema:
                $ref: '#/components/schemas/DefaultResponse'
          description: Invalid request format
      security:
      - sessionToken: []
      summary: Update business account
      tags:
      - business
  /business/menuImages/:
    post:
      description: This endpoint is used to add a new menu image to business details.
        Returns a new fileId to be used with '/file/' endpoints.
      operationId: addMenuImage
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostBusinessAccountMenuImageResponse'
          description: Successful operation
        "400":
          content:
            application/json:
              example:
                status: INVALID_REQUEST
              schema:
                $ref: '#/components/schemas/DefaultResponse'
          description: Invalid request format
      security:
      - sessionToken: []
      summary: Add menu image to business
      tags:
      - business
  /business/menuImages/{menuImageId}:
    delete:
      description: This endpoint is used to delete a menu image from business details.
      operationId: deleteMenuImage
      parameters:
      - description: Public id of the menu image to be deleted
        explode: false
        in: path
        name: menuImageId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponse'
          description: Successful operation
        "403":
          content:
            application/json:
              example:
                status: FORBIDDEN
              schema:
                $ref: '#/components/schemas/DefaultResponse'
          description: Forbidden
      security:
      - sessionToken: []
      summary: Delete menu image from business
      tags:
      - business
  /business/itemDefinitions:
    get:
      description: This endpoint is used to retrieve data about existing item definitions
        (benefits).
      operationId: getItemDefinitions
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetBusinessItemDefinitionsResponse'
          description: Successful operation
      security:
      - sessionToken: []
      summary: Get list of item definitions
      tags:
      - itemDefinitions
    post:
      description: This endpoint is used to add new item definitions (benefits).
      operationId: AddItemDefinition
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostBusinessItemDefinitionRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              example:
                status: CREATED
              schema:
                $ref: '#/components/schemas/PostBusinessItemDefinitionResponse'
          description: Successful operation
        "400":
          content:
            application/json:
              example:
                status: INVALID_REQUEST
              schema:
                $ref: '#/components/schemas/DefaultResponse'
          description: Invalid request format
      security:
      - sessionToken: []
      summary: Add a new item definition
      tags:
      - itemDefinitions
  /business/itemDefinitions/{definitionId}:
    delete:
      description: This endpoint is used to delete existing item definitions (benefits).
      operationId: deleteItemDefinition
      parameters:
      - description: Public id of the definition to update
        explode: false
        in: path
        name: definitionId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              example:
                status: OK
              schema:
                $ref: '#/components/schemas/DefaultResponse'
          description: Successful operation
        "400":
          content:
            application/json:
              example:
                status: INVALID_REQUEST
              schema:
                $ref: '#/components/schemas/DefaultResponse'
          description: Invalid request format
        "404":
          content:
            application/json:
              example:
                status: NOT_FOUND
              schema:
                $ref: '#/components/schemas/DefaultResponse'
          description: Does not exist or forbidden
      security:
      - sessionToken: []
      summary: Delete an exiting item definition
      tags:
      - itemDefinitions
    patch:
      description: This endpoint is used to change details of existing item definitions
        (benefits).
      operationId: updateItemDefinition
      parameters:
      - description: Public id of the definition to update
        explode: false
        in: path
        name: definitionId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchBusinessItemDefinitionRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              example:
                status: OK
              schema:
                $ref: '#/components/schemas/DefaultResponse'
          description: Successful operation
        "400":
          content:
            application/json:
              example:
                status: INVALID_REQUEST
              schema:
                $ref: '#/components/schemas/DefaultResponse'
          description: Invalid request format
        "404":
          content:
            application/json:
              example:
                status: NOT_FOUND
              schema:
                $ref: '#/components/schemas/DefaultResponse'
          description: Does not exist or forbidden
      security:
      - sessionToken: []
      summary: Update an exiting item definition
      tags:
      - itemDefinitions
...
components:
  schemas:
    DefaultResponse:
      example:
        message: message
        status: null
      properties:
        status:
          $ref: '#/components/schemas/DefaultResponseStatusEnum'
        message:
          type: string
      type: object
    DefaultResponseStatusEnum:
      enum:
      - OK
      - CREATED
      - NOT_FOUND
      - FORBIDDEN
      - ALREADY_EXISTS
      - UNAUTHORIZED
      - INVALID_REQUEST
      - UNKNOWN_ERROR
      - CONFLICT
      type: string
...
    PostBusinessAccountRequest:
      example:
        name: test business
        address: ul. Łojasiewicza 6 Kraków Polska
        gpsCoordinates: +48.8577+002.295/
        nip: 1234567899
        krs: 1234567890
        regon: 12345678901234
        ownerName: Jan Kowalski
      properties:
        name:
          type: string
          x-go-custom-tag: binding:"required"
        address:
          type: string
          x-go-custom-tag: binding:"required"
        gpsCoordinates:
          type: string
          x-go-custom-tag: binding:"required"
        description:
          type: string
          x-go-custom-tag: binding:"required"
        nip:
          type: string
          x-go-custom-tag: binding:"required"
        krs:
          type: string
          x-go-custom-tag: binding:"required"
        regon:
          type: string
          x-go-custom-tag: binding:"required"
        ownerName:
          type: string
          x-go-custom-tag: binding:"required"
      type: object
    PostBusinessAccountResponse:
      example:
        publicId: eGVlbGlCaWU5Z2FoCg
        bannerImageId: eGVlbGlCaWU5Z2FoCg
        iconImageId: eGVlbGlCaWU5Z2FoCg
      properties:
        publicId:
          type: string
        bannerImageId:
          type: string
        iconImageId:
          type: string
      type: object
    GetBusinessAccountResponse:
      allOf:
      - $ref: '#/components/schemas/PublicBusinessDetailsAPIModel'
      - $ref: '#/components/schemas/GetBusinessAccountResponse_allOf'
    PatchBusinessAccountRequest:
      example:
        name: test
      properties:
        name:
          type: string
        description:
          type: string
      type: object
    PostBusinessAccountMenuImageResponse:
      example:
        imageId: S2VlNWVlQ2Foajh1Cg
      properties:
        imageId:
          example: S2VlNWVlQ2Foajh1Cg
          type: string
      type: object
    PatchBusinessItemDefinitionRequest:
      allOf:
      - $ref: '#/components/schemas/ItemDefinitionAPIModel'
    PostBusinessItemDefinitionRequest:
      allOf:
      - $ref: '#/components/schemas/ItemDefinitionAPIModel'
    PostBusinessItemDefinitionResponse:
      example:
        publicId: ZmVlQ2ExYWlnaDRPCg
      properties:
        publicId:
          example: ZmVlQ2ExYWlnaDRPCg
          type: string
      type: object
    GetBusinessItemDefinitionsResponse:
      example:
        itemDefinitions:
        - publicId: QWl5ZWlwaG8ySWl2Cg
          name: Cool item
          price: 10
          description: idk
          imageId: QWhDOXBvdTdhaTZQCg
          startDate: 2023-04-21T15:25:37+02:00
          endDate: 2024-04-21T17:25:37+02:00
          maxAmount: 2
          available: true
        - publicId: QWl5ZWlwaG8ySWl2Cg
          name: Cool item
          price: 10
          description: idk
          imageId: QWhDOXBvdTdhaTZQCg
          startDate: 2023-04-21T15:25:37+02:00
          endDate: 2024-04-21T17:25:37+02:00
          maxAmount: 2
          available: true
      properties:
        itemDefinitions:
          items:
            $ref: '#/components/schemas/ItemDefinitionAPIModel'
          type: array
      type: object
...
    PublicBusinessDetailsAPIModel:
      example:
        publicId: ZUNoYWkyYWV4aVRhCg
        name: test
        address: ul. Łojasiewicza 6
        gpsCoordinates: +48.8577+002.295/
        bannerImageId: WG9ocGg2b0h1ZWNoCg
        iconImageId: c2VlbGVlNEVlV29vCg
        menuImageIds:
        - bXU5YWltMm1haUdpCg
        itemDefinitions:
        - publicId: QWl5ZWlwaG8ySWl2Cg
          name: Cool item
          price: 10
          description: idk
          imageId: QWhDOXBvdTdhaTZQCg
          startDate: 2023-04-21T15:25:37+02:00
          endDate: 2024-04-21T17:25:37+02:00
          maxAmount: 2
          available: true
      properties:
        publicId:
          type: string
        name:
          type: string
        address:
          type: string
        gpsCoordinates:
          type: string
        bannerImageId:
          type: string
        description:
          type: string
        iconImageId:
          type: string
        menuImageIds:
          items:
            type: string
          type: array
        itemDefinitions:
          items:
            $ref: '#/components/schemas/ItemDefinitionAPIModel'
          type: array
      type: object
    ItemDefinitionAPIModel:
      example:
        publicId: QWl5ZWlwaG8ySWl2Cg
        name: Cool item
        price: 10
        description: idk
        imageId: QWhDOXBvdTdhaTZQCg
        startDate: 2023-04-21T15:25:37+02:00
        endDate: 2024-04-21T17:25:37+02:00
        maxAmount: 2
        available: true
      properties:
        publicId:
          type: string
        name:
          type: string
        price:
          type: integer
        description:
          type: string
        imageId:
          type: string
        startDate:
          format: date-time
          nullable: true
          type: string
        endDate:
          format: date-time
          nullable: true
          type: string
        maxAmount:
          type: integer
        available:
          type: boolean
      type: object
...
    GetBusinessAccountResponse_allOf:
      properties:
        nip:
          type: string
        krs:
          type: string
        regon:
          type: string
        ownerName:
          type: string
      type: object
      example: null
...
  securitySchemes:
    sessionToken:
      bearerFormat: "Just the token string received from the API. In reality consists\
        \ of the token id and the actual token, separated by a colon."
      scheme: bearer
      type: http