# internal/api/models/model_get_business_account_response.go
/*
 * StampWallet API Server
 *
 * StampWallet API Server REST Specification
 *
 * API version: 0.1.0
 * Contact: fbstachura@gmail.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package api

type GetBusinessAccountResponse struct {
	PublicId string `json:"publicId,omitempty"`

	Name string `json:"name,omitempty"`

	Address string `json:"address,omitempty"`

	GpsCoordinates string `json:"gpsCoordinates,omitempty"`

	BannerImageId string `json:"bannerImageId,omitempty"`

	Description string `json:"description,omitempty"`

	IconImageId string `json:"iconImageId,omitempty"`

	MenuImageIds []string `json:"menuImageIds,omitempty"`

	ItemDefinitions []ItemDefinitionApiModel `json:"itemDefinitions,omitempty"`

	Nip string `json:"nip,omitempty"`

	Krs string `json:"krs,omitempty"`

	Regon string `json:"regon,omitempty"`

	OwnerName string `json:"ownerName,omitempty"`
}